+++ CHANGELOG.md
@@ -6,12 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added tests for new functionalities in SAPRFC and SAPRFCV2 regarding passing credentials
 - Added new params for mapping and reordering DataFrame for `Genesys` task and flow.
 
 ### Fixed
 
 ### Changed
-
+- Changed __init__ in SAPRFC and SAPRFCV2 class in source in order to raise warning in prefect when credentials will be taken from DEV.
 
 ## [0.4.22] - 2023-11-15
 ### Added